requirements.txt
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.1
openpyxl==3.11.0
pyyaml==6.0.1
fastparquet==2024.2.0
